+++ UVtools.GUI/Controls/Tools/CtrlToolLayerImport.Designer.cs
@@ -280,6 +280,7 @@
             // 
             this.AutoScaleDimensions = new System.Drawing.SizeF(9F, 20F);
             this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
+            this.AutoSizeMode = System.Windows.Forms.AutoSizeMode.GrowAndShrink;
             this.ButtonOkText = "Import";
             this.Controls.Add(this.cbDiscardRemainingLayers);
             this.Controls.Add(this.cbReplaceSubsequentLayers);